Allow running `install-testsuite-deps.py` reducing the needed imports, specially, `unittest2`.
